Agregar la nueva conexion a la BBDD
Cambie la connectionString de la base de datos del proyecto anterior, elimine las migraciones y cree un modelo y controlador de prueba
