+++ DSS_Scoring/appsettings.json
@@ -1,6 +1,6 @@
 {
   "ConnectionStrings": {
-    "SUPABASE_CONNECTION": "User Id=postgres.xxxxxxxxxxxx;Password=xxxxxxxxxxx;Server=aws-0-us-west-1.pooler.supabase.com;Port=5432;Database=postgres;"
+    "NEON_CONNECTION": "User Id=neondb_owner;Password=zmt3GXCKP6Er;Server=ep-lively-dawn-a5jirgb2.us-east-2.aws.neon.tech;Port=5432;Database=neondb;"
   },
   "Logging": {
     "LogLevel": {
